feat(opal): introduce RootLayout.App + RootLayout.MainContent; remove AppLayouts
- Adds `RootLayout.App` (flex-col wrapper, replaces old `MainContent`)
- Adds `RootLayout.MainContent` (flex-1 overflow-auto scrollable slot inside App)
- Extracts `AppHeader`, `AppFooter`, and `AppChrome` into `sections/app-chrome/`
  so app-chrome logic lives in the app, not Opal
- `AppChrome` wraps `RootLayout.App` with `@container`, `data-main-container`,
  and mouse-focus restoration handlers
- Deletes `web/src/layouts/app-layouts.tsx`; all six consumers migrated to the
  new layout hierarchy via `app/app/layout.tsx`
- Background overlay effects moved into `AppPage.tsx` (the only consumer)
- README and Storybook stories updated to reflect the new component hierarchy
