+++ web/lib/opal/src/layouts/root/README.md
@@ -3,8 +3,9 @@
 **Import:** `import { RootLayout } from "@opal/layouts";`
 
 Namespaced layout primitives for the app shell. Provides a full-viewport flex
-row with a controlled sidebar, optional permanent panels, and a main content
-area with optional pinned header/footer bars.
+row with a controlled sidebar, optional permanent panels, and an `App` column
+that contains a pinned header, a scrollable main content area, and a pinned
+footer.
 
 ## Components
 
@@ -14,7 +15,7 @@ Full-viewport flex row that wraps all other `RootLayout` primitives.
 
 | Prop | Type | Default | Description |
 |------|------|---------|-------------|
-| `children` | `ReactNode` | — | `Sidebar`, `LeftPanel`, `MainContent`, `RightPanel` |
+| `children` | `ReactNode` | — | `Sidebar`, `LeftPanel`, `App`, `RightPanel` |
 
 ### Sidebar
 
@@ -33,16 +34,21 @@ Controlled sidebar that handles three viewport sizes:
 | `onFoldToggle` | `() => void` | — | Called when the sidebar should toggle |
 | `children` | `ReactNode` | — | Sidebar shell and body content |
 
+### App
+
+`flex-1 flex-col` block that fills the remaining horizontal space between
+`Sidebar` and any panels. Use this as the direct child of `Root` to wrap
+`Header`, `MainContent`, and `Footer`. Accepts all standard `div` props
+(e.g. `className`, `data-*`, `onMouseDown`) for consumer-level overrides.
+
 ### MainContent
 
-`flex-1 flex-col` block that fills the remaining horizontal space. Direct children
-stack vertically, so `Header`, a scrollable content div (`flex-1 overflow-auto`),
-and `Footer` pin correctly without an extra wrapper. Accepts all standard `div`
-props (e.g. `className`, `data-*`) for consumer-level overrides.
+`flex-1 overflow-auto` scrollable slot inside `App`. Place page content here.
+Accepts all standard `div` props.
 
 ### LeftPanel / RightPanel
 
-Permanent `shrink-0` columns that push `MainContent` rather than overlaying it.
+Permanent `shrink-0` columns that push `App` rather than overlaying it.
 Width is caller-supplied via `className` (e.g. `className="w-80"`).
 
 | Prop | Type | Default | Description |
@@ -52,12 +58,11 @@ Width is caller-supplied via `className` (e.g. `className="w-80"`).
 
 ### Header
 
-Pinned `shrink-0` top bar inside `MainContent`. Use inside a `flex flex-col`
-content wrapper so it stays above the scrollable area.
+Pinned `shrink-0` top bar inside `App`.
 
 ### Footer
 
-Pinned `shrink-0` bottom bar inside `MainContent`.
+Pinned `shrink-0` bottom bar inside `App`.
 
 | Prop | Type | Default | Description |
 |------|------|---------|-------------|
@@ -92,9 +97,17 @@ import { RootLayout } from "@opal/layouts";
   <RootLayout.Sidebar folded={folded} onFoldToggle={toggle}>
     <MySidebarShell />
   </RootLayout.Sidebar>
-  <RootLayout.MainContent>
-    {children}
-  </RootLayout.MainContent>
+  <RootLayout.App>
+    <RootLayout.Header>
+      <AppHeader />
+    </RootLayout.Header>
+    <RootLayout.MainContent>
+      {children}
+    </RootLayout.MainContent>
+    <RootLayout.Footer>
+      <AppFooter />
+    </RootLayout.Footer>
+  </RootLayout.App>
 </RootLayout.Root>
 ```
 
@@ -108,32 +121,28 @@ import { RootLayout } from "@opal/layouts";
   <RootLayout.LeftPanel className="w-64">
     <FilterPanel />
   </RootLayout.LeftPanel>
-  <RootLayout.MainContent>
-    {children}
-  </RootLayout.MainContent>
+  <RootLayout.App>
+    <RootLayout.MainContent>
+      {children}
+    </RootLayout.MainContent>
+  </RootLayout.App>
   <RootLayout.RightPanel className="w-80">
     <DetailPanel />
   </RootLayout.RightPanel>
 </RootLayout.Root>
 ```
 
-### With header and footer
+### With consumer-level overrides on App
 
 ```tsx
-<RootLayout.Root>
-  <RootLayout.Sidebar folded={folded} onFoldToggle={toggle}>
-    <MySidebarShell />
-  </RootLayout.Sidebar>
-  <RootLayout.MainContent>
-    <RootLayout.Header>
-      <AppHeader />
-    </RootLayout.Header>
-    <div className="flex-1 overflow-auto">
-      {children}
-    </div>
-    <RootLayout.Footer extraPadding={!isActiveChat}>
-      <AppFooter />
-    </RootLayout.Footer>
-  </RootLayout.MainContent>
-</RootLayout.Root>
+<RootLayout.App
+  className="@container relative"
+  data-main-container
+  onMouseDown={handleMouseDown}
+  onMouseUp={handleMouseUp}
+>
+  <RootLayout.Header><AppHeader /></RootLayout.Header>
+  <RootLayout.MainContent>{children}</RootLayout.MainContent>
+  <RootLayout.Footer><AppFooter /></RootLayout.Footer>
+</RootLayout.App>
 ```
