+++ web/lib/opal/src/layouts/root/components.tsx
@@ -112,7 +112,21 @@ function RootLayoutSidebar({
 }
 
 // ---------------------------------------------------------------------------
-// MainContent — fills remaining flex space
+// App — fills remaining flex space; use as direct child of Root
+// ---------------------------------------------------------------------------
+
+type RootLayoutAppProps = React.HTMLAttributes<HTMLDivElement>;
+
+function RootLayoutApp({ children, className, ...props }: RootLayoutAppProps) {
+  return (
+    <div className={cn("opal-root-layout__app", className)} {...props}>
+      {children}
+    </div>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// MainContent — scrollable content slot inside App
 // ---------------------------------------------------------------------------
 
 type RootLayoutMainContentProps = React.HTMLAttributes<HTMLDivElement>;
@@ -196,6 +210,7 @@ function RootLayoutFooter({
 export {
   RootLayoutRoot as Root,
   RootLayoutSidebar as Sidebar,
+  RootLayoutApp as App,
   RootLayoutMainContent as MainContent,
   RootLayoutLeftPanel as LeftPanel,
   RootLayoutRightPanel as RightPanel,
