+++ web/lib/opal/src/layouts/root/styles.css
@@ -41,9 +41,14 @@
   @apply opacity-100 pointer-events-auto;
 }
 
-/* Main content — fills remaining flex space; flex-col so Header/Footer pin correctly */
-.opal-root-layout__main {
+/* App — fills remaining flex space; flex-col so Header/Footer pin correctly */
+.opal-root-layout__app {
   @apply flex-1 flex flex-col overflow-hidden h-full w-full;
+}
+
+/* Main content — scrollable slot inside App */
+.opal-root-layout__main {
+  @apply flex-1 overflow-auto;
 }
 
 /* Panel — permanent column (width is caller-supplied via className) */
