+++ web/src/app/app/agents/create/page.tsx
@@ -1,10 +1,5 @@
 import AgentEditorPage from "@/refresh-pages/AgentEditorPage";
-import * as AppLayouts from "@/layouts/app-layouts";
 
 export default async function Page() {
-  return (
-    <AppLayouts.Root>
-      <AgentEditorPage />
-    </AppLayouts.Root>
-  );
+  return <AgentEditorPage />;
 }
