+++ web/src/app/app/settings/layout.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { usePathname } from "next/navigation";
-import * as AppLayouts from "@/layouts/app-layouts";
 import { SettingsLayouts } from "@opal/layouts";
 import { SidebarTab } from "@opal/components";
 import { SvgSliders } from "@opal/icons";
@@ -23,55 +22,53 @@ export default function Layout({ children }: LayoutProps) {
   const showAccountsAccessTab = showPasswordSection || showTokensSection;
 
   return (
-    <AppLayouts.Root>
-      <SettingsLayouts.Root width="lg">
-        <SettingsLayouts.Header icon={SvgSliders} title="Settings" divider />
+    <SettingsLayouts.Root width="lg">
+      <SettingsLayouts.Header icon={SvgSliders} title="Settings" divider />
 
-        <SettingsLayouts.Body>
-          <Section
-            flexDirection="row"
-            justifyContent="start"
-            alignItems="start"
-            gap={1.5}
+      <SettingsLayouts.Body>
+        <Section
+          flexDirection="row"
+          justifyContent="start"
+          alignItems="start"
+          gap={1.5}
+        >
+          {/* Left: Tab Navigation */}
+          <div
+            data-testid="settings-left-tab-navigation"
+            className="flex flex-col px-2 min-w-50"
           >
-            {/* Left: Tab Navigation */}
-            <div
-              data-testid="settings-left-tab-navigation"
-              className="flex flex-col px-2 min-w-50"
+            <SidebarTab
+              href="/app/settings/general"
+              selected={pathname === "/app/settings/general"}
             >
+              General
+            </SidebarTab>
+            <SidebarTab
+              href="/app/settings/chat-preferences"
+              selected={pathname === "/app/settings/chat-preferences"}
+            >
+              Chat Preferences
+            </SidebarTab>
+            {showAccountsAccessTab && (
               <SidebarTab
-                href="/app/settings/general"
-                selected={pathname === "/app/settings/general"}
+                href="/app/settings/accounts-access"
+                selected={pathname === "/app/settings/accounts-access"}
               >
-                General
+                Accounts & Access
               </SidebarTab>
-              <SidebarTab
-                href="/app/settings/chat-preferences"
-                selected={pathname === "/app/settings/chat-preferences"}
-              >
-                Chat Preferences
-              </SidebarTab>
-              {showAccountsAccessTab && (
-                <SidebarTab
-                  href="/app/settings/accounts-access"
-                  selected={pathname === "/app/settings/accounts-access"}
-                >
-                  Accounts & Access
-                </SidebarTab>
-              )}
-              <SidebarTab
-                href="/app/settings/connectors"
-                selected={pathname === "/app/settings/connectors"}
-              >
-                Connectors
-              </SidebarTab>
-            </div>
+            )}
+            <SidebarTab
+              href="/app/settings/connectors"
+              selected={pathname === "/app/settings/connectors"}
+            >
+              Connectors
+            </SidebarTab>
+          </div>
 
-            {/* Right: Tab Content */}
-            {children}
-          </Section>
-        </SettingsLayouts.Body>
-      </SettingsLayouts.Root>
-    </AppLayouts.Root>
+          {/* Right: Tab Content */}
+          {children}
+        </Section>
+      </SettingsLayouts.Body>
+    </SettingsLayouts.Root>
   );
 }
