+++ web/src/app/app/shared/[chatId]/page.tsx
@@ -3,7 +3,6 @@ import { redirect } from "next/navigation";
 import type { Route } from "next";
 import { requireAuth } from "@/lib/auth/requireAuth";
 import SharedChatDisplay from "@/app/app/shared/[chatId]/SharedChatDisplay";
-import * as AppLayouts from "@/layouts/app-layouts";
 import { Agent } from "@/lib/agents/types";
 
 // This is used for rendering a persona in the shared chat display
@@ -62,9 +61,5 @@ export default async function Page(props: PageProps) {
     chatSession?.persona_id ?? 0
   );
 
-  return (
-    <AppLayouts.Root>
-      <SharedChatDisplay chatSession={chatSession} persona={persona} />
-    </AppLayouts.Root>
-  );
+  return <SharedChatDisplay chatSession={chatSession} persona={persona} />;
 }
